Use httpsnoop instead of a response wrapper (#155)
It turns out httpsnoop, useful in general for the fraught task of
wrapping ResponseWriter, already has a func for wrapping a
ResponseWriter and getting the bytes written to the body. So just use
that.

This adds an abstraction `Capture` for wrapping `http.ResponseWriter`,
and uses it to reimplement `LoggingResponseWriter`.

Ideally we would abandon `LoggingResponseWriter`, because being a
concrete type means it can't properly wrap `http.ResponseWriter` -- if
there are additional methods (like `ReadFrom()`) on the underlying
`http.ResponseWriter`, they will not be present in the
`LoggingResponseWriter`. I've provided it for backward compatibility
though, and given it some test coverage. For many purposes it doesn't
matter if it doesn't have the additional interfaces (and in any case,
this implementation behaves the same as the prior).

One subtlety is that there's no implementation of `Header()` for
either of `Capture` or `LoggingResponseWriter`. For `Capture`, you can
just use `Header()` on the underlying response writer. For
`LoggingResponseWriter`, `Header()` is provided by the embedded
`http.ResponseWriter`, which is the wrapped writer. In both cases, the
underlying writer's `Header()` is what we want, because that's the
value callers would get (and mutate) anyway.