Add Remote Reconciler (#160)
This is a cut down version of the reconciler that is intended to be used
by services that don't own the watched resource.  For example, the
compute service may want to trigger cluster deletion when the
corresonding project is deleted.  This basically emulates a compacted
message queue where you see events for create, update and delete, and
they are all triggered on restart so you don't miss anything.
Controllers will need to manage deletion semantics by adding finalizers
onto watched resources if they so require.

Also while here, sneak in a fix that scales reconciles to the number of
CPUs avaialble, as it can take a long time to process 1000s of resources
on service start.