Fix Conversion Functions
The short version is the deletion timestamp wasn't correctly propagated
for organization and project scoped APIs.  This is ultimately due to
OpenAPI not doing aggregation properly with idiomatic datatypes, forcing
us to manually copy things (or miss them out in this case).  To try to
be more reliable, use JSON an a intermediary to ensure things are
propagated, and guard with tests.