Upgrade API Improvements
First imbue the upgrade callback to accept a context and the core
options.  The upgrade stuff will need a context, so best propagate even
if it's a TODO for now, and the options gives access to the service's
namespace via the CLI flag.  The other upgrade is to get the client
scheme from the factory, so it is aware of the service's managed custom
resources.