Message Queue Abstraction
The previous implementation used a cut down version of the reconciler,
but this bring along with it a load of baggage and unsued receivers.
This attempts to create an interface, and base kubernetes
implementation, that is more akin to what we would get from a real
message queue implementation.  It's been tuned to not do any caching and
thus not waste memory.

This also adds in a generic hook into the reconciler logic that stops
any clean up from being performed until all uses of that parent resource
are removed (with the exception of the main one used to tombstone the
deletion until that's all completed successfully).

It also adds in a generic function to generate resource references from
custom resources in a standard fashion.