Ignore Foreground Delete Finalizer (#170)
Some services internally use Kubernetes foreground cascading deletion to
control when resources are deleted, as such it will implicitly add a
finalizer to the resource which will be removed once all owned resources
are cleaned up.  We need to ignore this from our set of resource
references so as to not be misleading.

We also generically handle the fireground deletion stuff rather than
having it repeated everywhere.