Treat Controller Options as Generics
The major bug bear with the existing solution that used interfaces is
that every comtroller with options needed to do a type assertion, to get
back from the abstract interface to a concrete type, and given the
constructor didn't return an error, you'd need to just ignore the
possible type assertion error and tolerate it with linter directives.

The one tiny drawback is that each controller factory now requires an
options struct for type inference, for which we have provided a null
default.