Add Saga Handler (#192)
The saga pattern descibes a saga -- a series of steps -- that may be
required for an API request.  It's well aware that some intermediate
step can go wrong, most likely quotas, and prior steps will need to be
compensated for to avoid resource leaks etc.  This implements that in a
nice generic way so it can be used by all complex handlers that do lots
of things at once.