Test timeout of timeout cache (#198)
* Test timeout of timeout cache

This adds a clock abstraction to the TimeoutCache, and uses it to
check that the cache is invalidated after the given timeout.

* Test that timeout cache resets on .Set()

This adds a check that after timing out, the cache will reset when a
value is Set() (i.e., it won't stay expired).

I factored out the value presen and absent checks, now they are used
in a few places.