Clean Up Error Handling (#209)
Makes OpenAPI the source of truth for types, so we don't get out of
sync.  Removes any oauth2 specific error types that are now local to the
identity service that actually uses them.  Starts to make the server
error all over the place a thing of the past by making it the default
behaviour and marking the function deprecated.  Backs all this up with
tests that ensure the behaviour of each handler so we can verify they
behave as specified.  This has also been tested with each child repo to
ensure backward compatibility.