Update CORS Handling (#210)
A recent change to the OpenAPI helpers removed an error handler that
returned an explicit 500 error.  What we really should be doing is
raising a 404 if the route cannot be found.  This codifies this
behaviour with tests, and while we are in the area, add a bunch more to
test the rest of it and boost code coverage/confidence.