WWW-Authenticate Headers
Ultimately we are implementing RFC9728 which is OIDC protected resource
metadata so a service that provides no authentication (e.g. MCP) will
get a list of authentication servers that can generate tokens and
perform token exchange.

To that end we need a new unauthorized handler that adds the
WWW-Authenticate header and also sets the metadata server to the local
service's endpoint.