Add API Error Propagation (#211)
At present we just pass this through as a raw error, which will
ultimately end up as a 500 to the user.  Instead we should unpack the
error so it can be propagated for root cause analysis.  I've backed this
up with tests and fixed the long standing crash when a gateway error
ocurrs.